New code

+++ labrab1/labrab1.1.py
@@ -1,4 +1,4 @@
-a = float(input())
-b = float(input())
+a = int(input())
+b = int(input())
 print('S =',a*b)
 print('P =',2*(a+b))
